requirements.txt
flask
cassandra-driver==3.25.0
bcrypt
requests
kafka-python
plotly
urllib3
pandas
uuid
geopandas